requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
scikit-learn>=1.3.0
jupyter>=1.0.0
tqdm>=4.65.0
pandas>=2.0.0
mtcnn>=0.1.1
